LayoutTests/fast/forms/datalist/datalist-textinput-keydown.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
</head>
<body>

<input id="fruit" list="fruits" type="text"/>
<datalist id="fruits">
    <option>Apple</option>
    <option>Orange</option>
    <option>Pear</option>
</datalist>

<script>

description('Test for keyboard operations for &lt;input type=text&gt; with list attribute');

var input = document.getElementById("fruit");

jsTestIsAsync = true;

(async () => {

await UIHelper.activateElement(input);
eventSender.keyDown("downArrow");
eventSender.keyDown("\r");
shouldBe('input.value', '"Apple"');

input.value = "";

await UIHelper.activateElement(input);
eventSender.keyDown("upArrow");
eventSender.keyDown("\r");
shouldBe('input.value', '"Pear"');

input.value = "Appl";

await UIHelper.activateElement(input);
eventSender.keyDown("\r");
shouldBe('input.value', '"Appl"');

const arrowKeysCycleThroughOptions = !(UIHelper.isIOSFamily() || UIHelper.isMac());

input.value = "";

await UIHelper.activateElement(input);
eventSender.keyDown("downArrow");
eventSender.keyDown("upArrow");
eventSender.keyDown("\r");
shouldBe('input.value', arrowKeysCycleThroughOptions ? '"Pear"' : '"Apple"');

input.value = "";

await UIHelper.activateElement(input);
eventSender.keyDown("upArrow");
eventSender.keyDown("downArrow");
eventSender.keyDown("\r");
shouldBe('input.value', arrowKeysCycleThroughOptions ? '"Apple"' : '"Pear"');

finishJSTest();
})();

</script>

</body>
</html>
